+++ supp/S01-CRAN-Build.R
@@ -6,7 +6,7 @@ devtools::document(
   pkg = here::here(),
   roclets = c('rd', 'collate', 'namespace')
 )
-install.packages(here::here(), repos=NULL)
+install.packages(here::here(), repos = NULL)
 
 ##############################@
 ## Build and copy vignettes ----
@@ -37,8 +37,8 @@ file.remove(here("doc"))
 
 ##############################@
 ## Build website ----
-unlink(here::here("docs"), recursive=TRUE, force=TRUE)
-pkgdown::build_site(pkg=here::here(), preview = FALSE)
+unlink(here::here("docs"), recursive = TRUE, force = TRUE)
+pkgdown::build_site(pkg = here::here(), preview = FALSE)
 
 ##############################@
 ## Build and check package ----
@@ -49,4 +49,4 @@ system(paste(
   sprintf("R CMD check --as-cran --no-build-vignettes TKCat_%s.tar.gz", pv),
   sep = " ; "
 ))
-install.packages(here(sprintf("../TKCat_%s.tar.gz", pv)), repos=NULL)
+install.packages(here(sprintf("../TKCat_%s.tar.gz", pv)), repos = NULL)